rename project lucene to backend-services
