+++ backend-services/pom.xml
@@ -8,10 +8,10 @@
 		<version>3.2.4</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>lucene.engine</groupId>
-	<artifactId>lucene</artifactId>
+	<groupId>backend.services</groupId>
+	<artifactId>backend-services</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>lucene</name>
+	<name>backend-services</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
 		<java.version>17</java.version>
