+++ backend-services/src/main/java/backend/services/BackendServicesApplication.java
@@ -1,14 +1,14 @@
-package lucene.engine.lucene;
+package backend.services;
 
-import lucene.engine.lucene.controller.DirectoryManager;
+import backend.services.controller.DirectoryManager;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class LuceneApplication {
+public class BackendServicesApplication {
 
     public static void main(String[] args) {
         DirectoryManager.ensureDirectoryExists("DDI-cache");
-        SpringApplication.run(LuceneApplication.class, args);
+        SpringApplication.run(BackendServicesApplication.class, args);
     }
 }
