+++ backend-services/src/main/java/backend/services/service/OpenFDAResponse.java
@@ -1,4 +1,4 @@
-package lucene.engine.lucene.service;
+package backend.services.service;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
