+++ backend-services/src/main/java/backend/services/service/OpenFDAService.java
@@ -1,4 +1,4 @@
-package lucene.engine.lucene.service;
+package backend.services.service;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
