--- lucene/src/test/java/lucene/engine/lucene/LuceneApplicationTests.java
@@ -1,13 +1,0 @@
-package lucene.engine.lucene;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LuceneApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
